some minor changes in the setup_cpu.sh

+++ setup_cpu.sh
@@ -1,29 +1,35 @@
-#!/bin/bash
+# create the python environment
+if [ -f cpu_env ]
+then
+	virtualenv -p python3 cpu_env
+	. cpu_env/bin/activate
+	pip install -r requirements_cpu.txt
+fi
 
-virtualenv -p python3 cpu_env
-. cpu_env/bin/activate
-pip install -r requirements_cpu.txt
+# Load the models
+if [ ! -f models ]
+then
+	mkdir models
+	wget -P "./models/" -N "pollithy.com/rnn_model_fake_data.h5"
+	wget -P "./models/" -N "pollithy.com/DEM_model.h5"
+fi
 
-
-# Load models
-mkdir models
-wget -P "./models/" -N "pollithy.com/rnn_model_fake_data.h5"
-wget -P "./models/" -N "pollithy.com/DEM_model.h5"
-
-
-# Load data
-mkdir data
-
-# wrong files
-# wget -P "./data/" -N "pollithy.com/DEM_Zylinder.csv"
-# wget -P "./data/" -N "pollithy.com/DEM_holzkugeln.csv"
-
-DataSetsArray=("Zylinder"  "Pfeffer"  "Kugeln"  "Weizen", "DEM_Cylinder_115")
-
-for datasetname in ${DataSetsArray[*]}; do
-     echo $datasetname
-     dataurl="pollithy.com/${datasetname}.zip"
-     zipname="${datasetname}.zip"
-     wget -P "./data/" -N "${dataurl}"
-     unzip -n "./data/${zipname}" -d "./data/"
-done
+# Load the data
+DataSetsArray=(
+	"Zylinder"
+	"Pfeffer"
+	"Kugeln"
+	"Weizen"
+	"DEM_Cylinder_115"
+)
+if [ ! -f data ]
+then
+	mkdir data
+	for datasetname in ${DataSetsArray[*]}; do
+	     echo $datasetname
+	     dataurl="pollithy.com/${datasetname}.zip"
+	     zipname="${datasetname}.zip"
+	     wget -P "./data/" -N "${dataurl}"
+	     unzip -n "./data/${zipname}" -d "./data/"
+	done
+fi
